Move code for getting envs at run time to configs folder
