+++ docker/config.js.tmpl
@@ -49,7 +49,7 @@ window._env_ = {
     "REACT_APP_SUPERSET_MDA_POINT_CHILD_REPORT_DATA_SLICE": "{{ getv "/react/app/superset/mda/point/child/report/data/slice" "578" }}",
     "REACT_APP_SUPERSET_MDA_POINT_REPORTING_JURISDICTIONS_FOCUS_AREA_LEVEL": "{{ getv "/react/app/superset/mda/point/reporting/jurisdictions/focus/area/level" "2" }}",
     "REACT_APP_ENABLE_OPENSRP_OAUTH": "{{ getv "/react/app/enable/opensrp/oauth" "true" }}",
-    "REACT_APP_OPENSRP_CLIENT_ID": "{{ getv "/react/app/opensrp/client/id" "reveal-stage-server" }}",
+    "REACT_APP_OPENSRP_CLIENT_ID": "{{ getv "/react/app/opensrp/client/id" "hunter2" }}",
     "REACT_APP_OPENSRP_ACCESS_TOKEN_URL": "{{ getv "/react/app/opensrp/access/token/url" "https://keycloak-stage.smartregister.org/auth/realms/reveal-stage/protocol/openid-connect/token" }}",
     "REACT_APP_OPENSRP_AUTHORIZATION_URL": "{{ getv "/react/app/opensrp/authorization/url" "https://keycloak-stage.smartregister.org/auth/realms/reveal-stage/protocol/openid-connect/auth" }}",
     "REACT_APP_OPENSRP_USER_URL": "{{ getv "/react/app/opensrp/user/url" "https://reveal-stage.smartregister.org/opensrp/user-details" }}",
@@ -57,7 +57,7 @@ window._env_ = {
     "REACT_APP_OPENSRP_OAUTH_STATE": "{{ getv "/react/app/opensrp/oauth/state" "opensrp" }}",
     "REACT_APP_ENABLE_ONADATA_OAUTH": "{{ getv "/react/app/enable/onadata/oauth" "false" }}",
     "REACT_APP_GISIDA_MAPBOX_TOKEN": "{{ getv "/react/app/gisida/mapbox/token" "pk.eyJ1IjoidGhhaWxhbmRidmJkIiwiYSI6ImNqdnh4b3d6NTBhNW4zeXBjc243OXpmamgifQ.5eHgZf6XwbUV8tvkmlOLZw" }}",
-    "REACT_APP_DIGITAL_GLOBE_CONNECT_ID": "{{ getv "/react/app/digital/globe/connect/id" "2293a3b8-1e33-4c16-92eb-0c842c4bc466" }}",
+    "REACT_APP_DIGITAL_GLOBE_CONNECT_ID": "{{ getv "/react/app/digital/globe/connect/id" "" }}",
     "REACT_APP_DATE_FORMAT": "{{ getv "/react/app/date/format" "yyyy-MM-dd" }}",
     "REACT_APP_DEFAULT_PLAN_DURATION_DAYS": "{{ getv "/react/app/default/plan/duration/days" "20" }}",
     "REACT_APP_PLAN_VERSION": "{{ getv "/react/app/plan/version" "1" }}",
