+++ src/helpers/tests/utils.test.tsx
@@ -48,7 +48,6 @@ import {
   PapaResult,
   removeNullJurisdictionPlans,
   roundToPrecision,
-  setEnv,
   SettingConfiguration,
 } from '../utils';
 import { refreshTokenResponse, userAuthData } from './fixtures';
@@ -565,14 +564,5 @@ describe('helpers/utils', () => {
     });
     expect(displayErrorSpy).toHaveBeenCalledWith(new Error(errorMessage));
     MockDate.reset();
-  });
-
-  it('sets an env var value to integer type', () => {
-    const envModule = require('../../configs/env');
-    envModule.SUPERSET_MAX_RECORDS = 1;
-    let recordsNumber = Number(setEnv('REACT_APP_SUPERSET_MAX_RECORDS', '1'));
-    expect(typeof recordsNumber).toEqual('number');
-    recordsNumber = setEnv('REACT_APP_SUPERSET_MAX_RECORDS', '1');
-    expect(typeof recordsNumber).toEqual('string');
   });
 });
